dev/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
av==9.0.0
base58==2.1.1
bitarray==3.0.0
certifi==2024.7.4
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
ckzg==2.0.1
constantly==23.10.4
cryptography==44.0.0
cytoolz==1.0.0
daphne==4.1.2
Django==4.2.13
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-filter==24.3
django-grip==3.5.1
djangorestframework==3.15.1
dnspython==2.7.0
drf-yasg==1.21.7
ecdsa==0.19.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
exceptiongroup==1.2.2
frozenlist==1.5.0
gevent==24.11.1
google-crc32c==1.6.0
greenlet==3.1.1
gripcontrol==4.2.0
h11==0.14.0
h2==4.1.0
hexbytes==1.2.1
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
ifaddr==0.2.0
importlib_metadata==7.1.0
incremental==24.7.2
inflection==0.5.1
MarkupSafe==3.0.2
mnemonic==0.21
MouseInfo==0.1.3
msgpack==1.1.0
mss==9.0.2
multidict==6.1.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==24.1
parsimonious==0.10.0
pefile==2023.2.7
pillow==10.4.0
priority==1.3.0
propcache==0.2.0
psutil==6.0.0
psycopg2-binary==2.9.9
pubcontrol==3.5.0
py-spy==0.3.14
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pyee==12.0.0
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.10.1
pylibsrtp==0.10.0
PyMsgBox==1.0.9
pyngrok==7.2.1
PyOpenGL==3.1.0
pyOpenSSL==24.3.0
pyperclip==1.9.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
PyScreeze==1.0.1
PySDL2==0.9.16
PySocks==1.7.1
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
pyunormalize==16.0.0
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.1
redis==5.2.0
regex==2024.11.6
requests==2.32.2
rlp==4.0.1
selenium==4.21.0
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tomli==2.1.0
toolz==1.0.0
trio==0.26.0
trio-websocket==0.11.1
Twisted==24.10.0
txaio==23.1.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
web3==7.6.0
websocket==0.2.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
whitenoise==6.8.2
wsproto==1.2.0
yarl==1.17.1
zipp==3.19.0
zope.event==5.0
zope.interface==7.1.1